refactor: implement route groups
