+++ apps/www/src/app/(authentication)/register/page.tsx
@@ -2,8 +2,8 @@ import Link from "next/link";
 import dynamic from "next/dynamic";
 import { getSession } from "@/lib/auth";
 import { redirect } from "next/navigation";
-import { V1Link } from "../components/v1-link";
 import { RegisterForm } from "./components/form";
+import { V1Link } from "@/app/components/v1-link";
 
 const BrowserWarning = dynamic(
   () => import("@umamin/ui/components/browser-warning"),
