+++ apps/www/src/app/(profile)/inbox/components/received/list.tsx
